requirements.txt
flask
flask-cors
sqlalchemy
pymysql
python-docx
python-pptx
bcrypt
